+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * StrUtil增加ordinalIndexOf方法（感谢@【四川】开心一笑）
 * 增加H2和SQLServer2012方言支持
 * 邮件支持发件人姓名（issue#IGMW2@Gitee）
+* db部分中Session增加默认数据源的Session
 
 ### Bug修复
 * 修复Oracle下Date类型字段带时分秒取出丢失问题（感谢@烟雨江南）
@@ -16,6 +17,7 @@
 * 邮件修复附件名乱码问题（issue#IGMW2@Gitee）
 * 修复 DateBetween.betweenYear中月份一致导致判断有误问题（issue#IGN0N@Gitee）
 * 修复JSONStrFormater在字符串中出现“[”、“{”等导致的换行错误问题（感谢@【北京】宁静）
+* 修复Props中两次创建Writer问题
 
 -------------------------------------------------------------------------------------------------------------